+++ src/schemathesis/specs/graphql/__init__.py
@@ -1,1 +1,1 @@
-from .loaders import from_dict, from_url, from_wsgi
+from .loaders import from_asgi, from_dict, from_url, from_wsgi