+++ test/apps/_graphql/__init__.py
@@ -1,1 +1,1 @@
-from . import _flask
+from . import _fastapi, _flask