+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 
 - Custom keyword arguments to ``schemathesis.graphql.from_url`` that are proxied to ``requests.post``.
 - ``from_wsgi`` loader for GraphQL apps. `#1097`_
+- ``from_asgi`` loader for GraphQL apps. `#1100`_
 - Support for ``app`` & ``base_url`` arguments for the ``from_pytest_fixture`` runner.
 
 **Changed**
@@ -1766,6 +1767,7 @@ Deprecated
 .. _0.3.0: https://github.com/schemathesis/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/schemathesis/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#1100: https://github.com/schemathesis/schemathesis/issues/1100
 .. _#1097: https://github.com/schemathesis/schemathesis/issues/1097
 .. _#1093: https://github.com/schemathesis/schemathesis/issues/1093
 .. _#1073: https://github.com/schemathesis/schemathesis/issues/1073